# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "AI Agent configuration platform with Skills and MCP support",
        "title": "Eino Agent API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@aiops.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/mcp/configs": {
            "get": {
                "description": "Get a list of all MCP server configurations",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mcp"
                ],
                "summary": "List all MCP configs",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_fisk086_taskmate_internal_schema.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_fisk086_taskmate_internal_schema.APIResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new MCP server configuration",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mcp"
                ],
                "summary": "Create a new MCP config",
                "parameters": [
                    {
                        "description": "MCP config data",
                        "name": "config",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_fisk086_taskmate_internal_schema.CreateMCPConfigRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_fisk086_taskmate_internal_schema.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_fisk086_taskmate_internal_schema.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_fisk086_taskmate_internal_schema.APIResponse"
                        }
                    }
                }
            }
        },
        "/mcp/configs/{id}": {
            "put": {
                "description": "Update an existing MCP server configuration",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mcp"
                ],
                "summary": "Update an MCP config",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "MCP Config ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "MCP config data",
                        "name": "config",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_fisk086_taskmate_internal_schema.CreateMCPConfigRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_fisk086_taskmate_internal_schema.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_fisk086_taskmate_internal_schema.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_fisk086_taskmate_internal_schema.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_fisk086_taskmate_internal_schema.APIResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete an MCP server configuration",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mcp"
                ],
                "summary": "Delete an MCP config",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "MCP Config ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_fisk086_taskmate_internal_schema.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_fisk086_taskmate_internal_schema.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_fisk086_taskmate_internal_schema.APIResponse"
                        }
                    }
                }
            }
        },
        "/mcp/configs/{id}/sync": {
            "post": {
                "description": "Sync tools from an MCP server",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mcp"
                ],
                "summary": "Sync MCP server tools",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "MCP Config ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Sync request",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_fisk086_taskmate_internal_schema.SyncMCPServerRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_fisk086_taskmate_internal_schema.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_fisk086_taskmate_internal_schema.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_fisk086_taskmate_internal_schema.APIResponse"
                        }
                    }
                }
            }
        },
        "/mcp/configs/{id}/tools": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mcp"
                ],
                "summary": "List tools synced for an MCP config",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "MCP Config ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_fisk086_taskmate_internal_schema.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_fisk086_taskmate_internal_schema.APIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_fisk086_taskmate_internal_schema.APIResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
        "github_com_fisk086_taskmate_internal_schema.CreateMCPConfigRequest": {
            "type": "object",
            "required": [
                "key",
                "name"
            ],
            "properties": {
                "config": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "endpoint": {
                    "type": "string"
                },
                "key": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "transport": {
                    "type": "string"
                }
            }
        },
...
        "github_com_fisk086_taskmate_internal_schema.MCPServer": {
            "type": "object",
            "properties": {
                "config_id": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "display_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "input_schema": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "is_active": {
                    "type": "boolean"
                },
                "tool_name": {
                    "type": "string"
                }
            }
        },
...
        "github_com_fisk086_taskmate_internal_schema.SyncMCPServerRequest": {
            "type": "object",
            "properties": {
                "create_capability": {
                    "type": "boolean"
                },
                "tools": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_fisk086_taskmate_internal_schema.MCPServer"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT access token (Authorization: Bearer plus space plus token)",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}